WIP: decouple AAD authentication from RestApi user
