+++ tests/unit/test_grizzly/auth/test_aad.py
@@ -1,0 +1,2 @@
+class TestAAD:
+    pass
