+++ tests/unit/test_grizzly/users/test_restapi.py
@@ -18,7 +18,7 @@ from _pytest.logging import LogCaptureFixture
 from requests.models import Response
 from json import dumps as jsondumps
 
-from grizzly.users.restapi import AuthMethod, RestApiUser, refresh_token
+from grizzly.users.restapi import RestApiUser
 from grizzly.users.base import AsyncRequests, RequestLogger, ResponseHandler, GrizzlyUser
 from grizzly.clients import ResponseEventSession
 from grizzly.types import GrizzlyResponse, RequestMethod, GrizzlyResponseContextManager
@@ -29,198 +29,11 @@ from grizzly.testdata.utils import transform
 from grizzly.exceptions import RestartScenario
 from grizzly_extras.transformer import TransformerContentType
 
-from tests.fixtures import GrizzlyFixture, ResponseContextManagerFixture
+from tests.fixtures import ResponseContextManagerFixture
 from tests.helpers import RequestSilentFailureEvent, RequestEvent, ResultSuccess
 
 
 RestApiScenarioFixture = Tuple[RestApiUser, GrizzlyContextScenario]
-
-
-@pytest.mark.usefixtures('grizzly_fixture')
-@pytest.fixture
-def restapi_user(grizzly_fixture: GrizzlyFixture) -> RestApiScenarioFixture:
-    scenario = GrizzlyContextScenario(1)
-    scenario.name = 'TestScenario'
-    scenario.context['host'] = 'test'
-    scenario.user.class_name = 'RestApiUser'
-
-    _, user, _ = grizzly_fixture('http://example.net', RestApiUser)
-
-    request = grizzly_fixture.request_task.request
-
-    scenario.tasks.add(request)
-
-    return cast(RestApiUser, user), scenario
-
-
-def test_refresh_token_client(restapi_user: RestApiScenarioFixture, mocker: MockerFixture) -> None:
-    [user, scenario] = restapi_user
-    decorator = refresh_token()
-
-    auth_context = user.context()['auth']
-
-    assert auth_context['refresh_time'] == 3000
-
-    class NotRefreshed(Exception):
-        pass
-
-    class Refreshed(Exception):
-        pass
-
-    def request(self: 'RestApiUser', *args: Tuple[Any, ...], **kwargs: Dict[str, Any]) -> None:
-        raise NotRefreshed()
-
-    mocker.patch(
-        'grizzly.users.restapi.RestApiUser.request',
-        decorator(request),
-    )
-
-    def get_token(self: 'RestApiUser', auth_method: AuthMethod) -> None:
-        if auth_method == AuthMethod.CLIENT:
-            raise Refreshed()
-
-    mocker.patch(
-        'grizzly.users.restapi.RestApiUser.get_token',
-        get_token,
-    )
-
-    try:
-        auth_client_context = auth_context['client']
-        request_task = RequestTask(RequestMethod.POST, name='test-request', endpoint='/api/test')
-        scenario.tasks.clear()
-        scenario.tasks.add(request_task)
-
-        # no authentication for api, request will be called which raises NotRefreshed
-        assert auth_context['provider'] is None
-        assert auth_client_context['id'] is None
-        assert auth_client_context['secret'] is None
-        assert auth_client_context['resource'] is None
-
-        with pytest.raises(NotRefreshed):
-            user.request(request_task)
-
-        auth_client_context['id'] = 'asdf'
-        auth_client_context['secret'] = 'asdf'
-        auth_context['refresh_time'] = 3000
-        auth_context['provider'] = 'http://login.example.com/oauth2'
-
-        # session has not started
-        with pytest.raises(NotRefreshed):
-            user.request(request_task)
-
-        user.session_started = time()
-        user.headers['Authorization'] = None
-
-        # session is fresh, but no token set (first call)
-        with pytest.raises(Refreshed):
-            user.request(request_task)
-
-        # token is fresh and set, no refresh
-        user.session_started = time()
-        user.headers['Authorization'] = 'Bearer asdf'
-
-        with pytest.raises(NotRefreshed):
-            user.request(request_task)
-
-        # authorization is set, but it is time to refresh token
-        user.session_started = time() - (cast(int, auth_context['refresh_time']) + 1)
-
-        with pytest.raises(Refreshed):
-            user.request(request_task)
-    finally:
-        pass
-
-
-def test_refresh_token_user(restapi_user: RestApiScenarioFixture, mocker: MockerFixture) -> None:
-    [user, scenario] = restapi_user
-    decorator = refresh_token()
-
-    auth_context = user.context()['auth']
-
-    assert auth_context['refresh_time'] == 3000
-
-    class NotRefreshed(Exception):
-        pass
-
-    class Refreshed(Exception):
-        pass
-
-    def request(self: 'RestApiUser', *args: Tuple[Any, ...], **kwargs: Dict[str, Any]) -> None:
-        raise NotRefreshed()
-
-    mocker.patch(
-        'grizzly.users.restapi.RestApiUser.request',
-        decorator(request),
-    )
-
-    def get_token(self: 'RestApiUser', auth_method: AuthMethod) -> None:
-        if auth_method == AuthMethod.USER:
-            raise Refreshed()
-
-    mocker.patch(
-        'grizzly.users.restapi.RestApiUser.get_token',
-        get_token,
-    )
-
-    try:
-        auth_user_context = auth_context['user']
-        request_task = RequestTask(RequestMethod.POST, name='test-request', endpoint='/api/test')
-        scenario.tasks.clear()
-        scenario.tasks.add(request_task)
-
-        # no authentication for api, request will be called which raises NotRefreshed
-        assert auth_user_context['username'] is None
-        assert auth_user_context['password'] is None
-        assert auth_user_context['redirect_uri'] is None
-        assert auth_context['provider'] is None
-
-        with pytest.raises(NotRefreshed):
-            user.request(request_task)
-
-        auth_context['client']['id'] = 'asdf'
-        auth_user_context['username'] = 'bob@example.com'
-        auth_user_context['password'] = 'HemligaArne'
-        auth_user_context['redirect_uri'] = '/authenticated'
-        auth_context['refresh_time'] = 3000
-        auth_context['provider'] = 'https://login.example.com/oauth2'
-
-        # session has not started
-        with pytest.raises(NotRefreshed):
-            user.request(request_task)
-
-        user.session_started = time()
-        user.headers['Authorization'] = None
-
-        # session is fresh, but no token set (first call)
-        with pytest.raises(Refreshed):
-            user.request(request_task)
-
-        # token is fresh and set, no refresh
-        user.session_started = time()
-        user.headers['Authorization'] = 'Bearer asdf'
-
-        with pytest.raises(NotRefreshed):
-            user.request(request_task)
-
-        # authorization is set, but it is time to refresh token
-        user.session_started = time() - (cast(int, auth_context['refresh_time']) + 1)
-
-        with pytest.raises(Refreshed):
-            user.request(request_task)
-
-        user.add_context({'auth': {'user': {'username': 'alice@example.com'}}})
-        assert user.headers.get('Authorization', '') is None
-        assert user._context['auth']['user'] == {
-            'username': 'alice@example.com',
-            'password': 'HemligaArne',
-            'redirect_uri': '/authenticated',
-        }
-
-        # new user in context, needs to get a new token
-        with pytest.raises(Refreshed):
-            user.request(request_task)
-    finally:
-        pass
 
 
 class TestRestApiUser:
